add build action, move language tools to seperate file
closes #2
